requirements.txt
Flask==2.3.3
torch==2.0.1
torchvision==0.15.2
ftfy==6.1.1
regex==2023.8.8
tqdm==4.66.1
Pillow==10.0.1
numpy==1.24.3
clip-by-openai==1.0
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2